backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
sqlalchemy[asyncio]==2.0.*
alembic==1.14.*
asyncpg==0.30.*
psycopg2-binary==2.9.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
anthropic==0.40.*
openai==1.58.*
PyMuPDF==1.25.*
python-docx==1.1.*
Pillow==11.*
pytesseract==0.3.*
python-multipart==0.0.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
apscheduler==3.10.*
slowapi==0.1.*
